Fix Actor coordinate handling for combatHandler
This is confusing -- one would think to use getX() and getY(), but that
doesn't work as expected. Those return the location of the actor within
the table, which we would then attempt to transform to stage
coordinates.
What we need to do is determine where a point is within the current
actor.
this.localToStageCoordinates(new Vector2(0, 0) is the bottom left corner
of the actor
this.localToStageCoordinates(new Vector2(0, getHeight()) is the top left
corner of the actor